refactor(profile): move to authroute
